refactor: flatten app/drawers and app/dialogs into feature folders
Group by feature instead of UI primitive: app/menu, app/settings, and
app/onboarding replace app/drawers and app/dialogs.
