+++ src/app/settings/ai-settings.tsx
